+++ braindecode/models/eeginception.py
@@ -108,6 +108,7 @@ class EEGInception(EEGModuleMixin, nn.Sequential):
             in_channels=None,
             n_classes=None,
             input_window_samples=None,
+            add_log_softmax=True,
     ):
         n_chans, n_outputs, n_times, = deprecated_args(
             self,
@@ -122,6 +123,7 @@ class EEGInception(EEGModuleMixin, nn.Sequential):
             n_times=n_times,
             input_window_seconds=input_window_seconds,
             sfreq=sfreq,
+            add_log_softmax=add_log_softmax,
         )
         del n_outputs, n_chans, chs_info, n_times, input_window_seconds, sfreq
         del in_channels, n_classes, input_window_samples
@@ -250,7 +252,7 @@ class EEGInception(EEGModuleMixin, nn.Sequential):
                 spatial_dim_last_layer * n_channels_last_layer,
                 self.n_outputs
             ),
-            nn.Softmax(1)
+            nn.LogSoftmax(dim=1) if self.add_log_softmax else nn.Identity(),
         ))
 
         _glorot_weight_zero_bias(self)