+++ braindecode/models/sleep_stager_blanco_2020.py
@@ -64,6 +64,7 @@ class SleepStagerBlanco2020(EEGModuleMixin, nn.Module):
             n_channels=None,
             n_classes=None,
             input_size_s=None,
+            add_log_softmax=True,
     ):
         n_chans, n_outputs, input_window_seconds, = deprecated_args(
             self,
@@ -78,6 +79,7 @@ class SleepStagerBlanco2020(EEGModuleMixin, nn.Module):
             n_times=n_times,
             input_window_seconds=input_window_seconds,
             sfreq=sfreq,
+            add_log_softmax=add_log_softmax,
         )
         del n_outputs, n_chans, chs_info, n_times, input_window_seconds, sfreq
         del n_channels, n_classes, input_size_s
@@ -121,7 +123,7 @@ class SleepStagerBlanco2020(EEGModuleMixin, nn.Module):
             self.fc = nn.Sequential(
                 nn.Dropout(dropout),
                 nn.Linear(self.len_last_layer, self.n_outputs),
-                nn.Softmax(dim=1)
+                nn.LogSoftmax(dim=1) if self.add_log_softmax else nn.Identity()
             )
 
     def _len_last_layer(self, n_channels, input_size):